+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.88.3](https://github.com/sinameraji/kimiflare/compare/v0.88.2...v0.88.3) (2026-06-15)
+
+
+### Bug Fixes
+
+* **agent:** preserve original plan across /fresh in plan mode ([#589](https://github.com/sinameraji/kimiflare/issues/589)) ([ed14cc6](https://github.com/sinameraji/kimiflare/commit/ed14cc612f60a7bca0a591922ccadacce386ed32))
+
+## [0.88.2](https://github.com/sinameraji/kimiflare/compare/v0.88.1...v0.88.2) (2026-06-15)
+
+
+### Bug Fixes
+
+* **memory:** route embeddings through AI Gateway /compat/embeddings ([#587](https://github.com/sinameraji/kimiflare/issues/587)) ([4648b9b](https://github.com/sinameraji/kimiflare/commit/4648b9b4def647299647878dae5cdbe5d5d4afbd))
+
 ## [0.88.1](https://github.com/sinameraji/kimiflare/compare/v0.88.0...v0.88.1) (2026-06-13)